añadi validaciones para la parte de productos

+++ src/main/java/com/example/onix/models/services/ProductService.java
@@ -1,5 +1,7 @@
 package com.example.onix.models.services;
 
+import com.example.onix.exceptions.NotFoundException;
+import com.example.onix.mappers.ProductMapper;
 import com.example.onix.models.dto.CommentDto;
 import com.example.onix.models.dto.ProductDto;
 import com.example.onix.models.entities.Product;
@@ -17,45 +19,54 @@ import java.util.Optional;
 public class ProductService implements IProductService {
 
     private final ProductRepository productoRepository;
+    private final ProductMapper productMapper;
 
     @Override
     public List<ProductDto> getAllProducts() {
-        return (List<ProductDto>) productoRepository.findAllDto();
+        List<Product> products = (List<Product>) productoRepository.findAll();
+        if (products.isEmpty()){
+            throw new NotFoundException("No se encontraron productos");
+        }
+        return productMapper.toDtoList(products);
     }
 
     @Override
-    public void saveProduct(Product product) {
-        productoRepository.save(product);
+    public ProductDto saveProduct(Product product) {
+       return productMapper.toDto(productoRepository.save(product));
     }
 
     @Override
     @Transactional
-    public void updateProduct(Long id, ProductDto product) {
+    public ProductDto updateProduct(Long id, ProductDto product) {
         Product existingProduct = productoRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Producto no encontrado con id: " + id));
+                .orElseThrow(() -> new NotFoundException("Producto no encontrado con id: " + id));
         existingProduct.setName(product.getName());
         existingProduct.setDescription(product.getDescription());
         existingProduct.setImage(product.getImage());
         existingProduct.setPrice(product.getPrice());
 
+        return productMapper.toDto(existingProduct);
     }
     @Override
     @Transactional
     public void updateComment(Long id, String comment){
         Product existingProduct = productoRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Producto no encontrado con id: "+ id));
+                .orElseThrow(()-> new NotFoundException("Producto no encontrado con id: "+ id));
         if (comment!= null && !comment.trim().isEmpty()) {
             existingProduct.getComments().add(comment);
         }
     }
     @Override
-    public Product getProductById(Long id) {
-        return productoRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Producto no encontrado con id: " + id));
+    public ProductDto getProductById(Long id) {
+        return productMapper.toDto(productoRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Producto no encontrado con id: " + id)));
     }
 
     @Override
     public void deleteProductById(Long id) {
+        if (!productoRepository.existsById(id)) {
+            throw new NotFoundException("Producto no encontrado con id: " + id);
+        }
         productoRepository.deleteById(id);
     }
 }
